added APIs

+++ projects/src/main/java/com/debjyotipaul/forms/CategoryCountForm.java
@@ -16,9 +16,9 @@ public class CategoryCountForm {
     String name;
 
     @JsonProperty
-    int value;
+    double value;
 
-    public CategoryCountForm(String name, int value) {
+    public CategoryCountForm(String name, double value) {
         this.name = name;
         this.value = value;
     }
@@ -31,11 +31,11 @@ public class CategoryCountForm {
         this.name = name;
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
 }
